requirements.txt
fastapi
uvicorn[standard]
pydantic[email]>=2.0
python-multipart
python-docx
PyPDF2
requests
beautifulsoup4
SQLAlchemy
psycopg2-binary
alembic
python-dotenv
anthropic
google-generativeai
google-cloud-texttospeech
openai-whisper
pydub
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]
pgvector
sentence-transformers
# py-hanspell from a fixed git repository
git+https://github.com/Hyunwoo2267/hanspell.git